==> 0_preprocessing/02_interpret_transform.ipynb <==
# Interpret and Transform

import pandas as pd

df_collision_original = pd.read_csv("./data/collisions1.csv", date_format="%d/%m/%Y", parse_dates=["date"])
df_collision_original.head().T

## 1. Make Categorical variables readable

The original dataset encodes its categorical variables as numbers. For ease of use, they should be converted into readable strings. 

There is a [Road Safety Open Dataset Guide](https://data.dft.gov.uk/road-accidents-safety-data/dft-road-casualty-statistics-road-safety-open-dataset-data-guide-2023.xlsx) that provides the key for these values and can be used as a mapper

df_interp = pd.read_excel("./data/dft-road-casualty-statistics-road-safety-open-dataset-data-guide-2023.xlsx")

df_collision_key = df_interp[df_interp["table"] == "accident"]

categorical_columns = ["police_force","accident_severity",
                       "day_of_week", "first_road_class","road_type","junction_detail",
                       "pedestrian_crossing_human_control","pedestrian_crossing_physical_facilities",
                       "light_conditions", "weather_conditions", "road_surface_conditions",
                       "special_conditions_at_site", "carriageway_hazards","urban_or_rural_area",
                       "trunk_road_flag"]

categorical_columns_mapping = {}

for col in categorical_columns:
    filtered_key = df_collision_key[df_collision_key["field name"] == col]
    categorical_columns_mapping[col] = pd.Series(filtered_key["label"].values,index=filtered_key["code/format"]).to_dict()

for col in categorical_columns:
    col_mapper = categorical_columns_mapping[col]
    df_collision_original[col] = df_collision_original[col].map(col_mapper)

df_collision_original.head().T

## 2. Filter for data collected by police officers who attended the scene

`did_police_officer_attend_scene_of_accident` tells us wether a police officer attended the scene of the accident.

The majority of the data (~70%) was collected by police officers who attended the scene of the accident. 

10% was collected over the counter at a police station and around 20% was collected from forms completed by members of the public ("self reporting" forms). It is acknowledged in the advice on completing the form that data from forms completed by members of the public may contain missing fields.

It's reasonable to assume too that police officers are likely to be more experienced with collecting this data making the data more reliable.

To reduce the number of missing fields and in an attempt to make the data more reliable, I have filtered the collision data by this field.

df_collision_original["did_police_officer_attend_scene_of_accident"].value_counts(normalize=True)

df_collision = df_collision_original[df_collision_original["did_police_officer_attend_scene_of_accident"] == 1].copy()
df_collision["did_police_officer_attend_scene_of_accident"].value_counts()

del df_collision["did_police_officer_attend_scene_of_accident"]

Convert categorical column numbers into strings

len(df_collision)

## 3. Create the Target Variable

`accident_severity` has three possible values: `Slight`, `Serious` and `Fatal`.

- Fatal
This covers casualties who sustained injuries which caused death less than 30 days after the accident.

- Serious
"Injuries classed as serious include: fractures, concussion, internal injuries, crushings, burns (excluding friction burns), severe cuts, severe general shock requiring medical treatment and injuries causing death 30 or more days after the accident." [Reported road casualty statistics: Background Quality Report](https://www.gov.uk/government/publications/reported-road-casualty-statistics-background-quality-report/reported-road-casualty-statistics-background-quality-report)

- Slight
"An injured casualty that is not classified as seriously injured, having an injury of a minor character such as a sprain (including neck whiplash injury), bruise or cut which are not judged to be severe, or slight shock requiring roadside attention. This definition includes injuries not requiring medical treatment." [Reported road casualty statistics: Background Quality Report](https://www.gov.uk/government/publications/reported-road-casualty-statistics-background-quality-report/reported-road-casualty-statistics-background-quality-report) 

**Assumption**: We assume that the severity categorisation is accurate. 

In this case we will categorise every collision in this dataset with a category of "Serious" or "Fatal" as being "severe".

df_collision["accident_severity"].value_counts()

df_collision["is_severe"] = df_collision["accident_severity"].isin(["Serious", "Fatal"]).astype(int)
del df_collision["accident_severity"]

df_collision["is_severe"].value_counts(normalize=True)

## 4. Handle date and time

The `date` column for humans captures information about year, month and day number and we can understand the differences and perhaps see patterns across years, months and days but for an algorithm to pick this up and see temporal differences, we often need to be more explicit and extract these variables.

Our data has already been broken down a little bit: we have separate columns for time and day of the week. 

We will include the following date and time variables and drop the date column:

- `time` (rounded to nearest hour)
- `day_of_week`
- `month` (extracted from date)
- `day_of_year` (extracted from date)

### Rounding time to the nearest hour

`time` records the the time of the accident.

There are 1440 unique values in our dataframe. It looks like a considerable number have been rounded to the nearest hour but not all. 

We will convert all of the time data to round to the nearest hour so that we can better see patterns.

The time category now has 24 unique values i.e. every hour block is present, so our category is complete. 


len(df_collision["time"].unique())

df_collision["time"].value_counts().head(5)

df_collision["time"].tail()

df_collision["time"] = pd.to_datetime(df_collision["time"], format="%H:%M").dt.round('60min').dt.strftime("%H:%M")

len(df_collision["time"].unique())

### Extracting month

We have the `date` field. It might be interesting to see if the month has a role in predicting severity.

df_collision["month"] = pd.to_datetime(df_collision["date"], format="mixed").dt.month_name()
df_collision["month"] = df_collision["month"].astype("category") 
category_order = ['January','February','March','April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December']
df_collision["month"] = df_collision["month"].cat.set_categories(category_order, ordered=True)
df_collision.head()

### Extract day of year from date and drop date column

To allow the algorithm to capture relative differences between dates, we will convert our date to number of days since January 1st 2023 i.e. we will use pandas `dayofyear` attribute which returns the day of the year on which the particular date occurs. This produces a continuous numerical value from 1 to 365, where 1 is the 1st day of the year i.e. 1st January 2023.

This new feature should be able to implicitly capture week based patterns so we have no real strong need for adding a week_of_year category.


df_collision["date"] = pd.to_datetime(df_collision["date"])
df_collision["day_of_year"] = df_collision["date"].dt.dayofyear
del df_collision["date"]

## 5. Create consistency

Our original categorical data consists of strings that are often capitalized. To maintain a degres of consistency, we replace all spaces and hyphens with underscores.

categorical_cols = list(df_collision.dtypes[(df_collision.dtypes == "object") | (df_collision.dtypes == "category")].index)

for col in categorical_cols:
    df_collision[col] = df_collision[col].str.strip()
    df_collision[col] = df_collision[col].str.lower().str.replace(' ','_')
    df_collision[col] = df_collision[col].str.replace('-','_')
    df_collision[col] = df_collision[col].str.replace('+','')

df_collision.tail().T

## Save data

df_collision.to_csv("./data/collisions2.csv", index=False)
