==> 0_preprocessing/03_remove_casualty_number.ipynb <==
# Remove casualty number

import pandas as pd

## Aim

I want to use the [Road Safety Data](https://www.data.gov.uk/dataset/cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data) to predict whether an accident is very serious i.e. likely to need an ambulance or other emergency services. The problem I am solving is centred around an imaginary situation and scenario.

**Imaginary scenario** 

We have developed a system that can detect when a collision involving casualties has happened on the road. The system uses cameras and sensors to feed this system information about the location, the environmental conditions and any particular hazards that exist on the road at the time. 

We are interested in ensuring the relevant emergency services are contacted immediately: if the accident is very severe, then an ambulance may be required.

To solve this problem we want to develop a way of predicting if a collision involving casualties is severe enough to warrant an ambulance.

**Solution**

To predict if a collision involving casualties is severe, I will implement a binary classification algorithm using the computed variable `is_severe` as the target variable.

In our imaginary scenario, there are cameras and sensors to feed this system information about the location, the environmental conditions and any particular hazards that exist on the road at the time. 

We assume that the system does not know the number of casualties so we will delete this information from the dataset. 

df_collision = pd.read_csv("./data/collisions2.csv")

del df_collision["number_of_casualties"]
df_collision.tail().transpose()

## Save data

df_collision.to_csv("./data/collisions3.csv", index=False)
